move WaitForDBAGrants into tabletserver package
